requirements.txt
Brotli==1.0.9
cachetools==5.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.1
httplib2==0.20.4
idna==3.3
mutagen==1.45.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.14.1
pyparsing==3.0.9
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.9
websockets==10.3
youtube-dl==2021.12.17
yt-dlp==2022.4.8
